Problems Resolved:
- Fixed couple areas where NPE might occur due to command parsing
- Optimized command execution performance

Features Added:
- TMSA commands can be run from the console

+++ bukkit/src/main/java/world/naturecraft/townymission/listeners/mission/money/EssentialMoneyListener.java
@@ -65,6 +65,6 @@ public class EssentialMoneyListener extends MissionListener {
             if (!localCheck.check()) return;
         }
 
-        doLogic(checker, MissionType.MONEY, player, (int) (event.getNewBalance().subtract(event.getOldBalance())).intValue());
+        doLogic(checker, MissionType.MONEY, player, (event.getNewBalance().subtract(event.getOldBalance())).intValue());
     }
 }
